Release/2.2.1

+++ src/main/java/io/github/shamrice/discapp/web/controller/abuse/AbuseController.java
@@ -19,6 +19,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import static io.github.shamrice.discapp.web.define.url.AbuseUrl.*;
@@ -43,13 +44,13 @@ public class AbuseController {
     private ApplicationService applicationService;
 
     @GetMapping(ABUSE_DELETE)
-    public ModelAndView getAbuseView(@RequestParam(name = "abuseId") Long abuseId,
-                                     @RequestParam(name = "discId", required = false) Long appId,
-                                     @RequestParam(name = "submitter", required = false) String submitter,
-                                     @RequestParam(name = "email", required = false) String email,
-                                     @RequestParam(name = "ip", required = false) String ip,
-                                     @RequestParam(name = "subject", required = false) String subject,
-                                     @RequestParam(name = "body", required = false) String body,
+    public ModelAndView getAbuseView(@RequestParam(name = ABUSE_ID_QUERY_PARAM) Long abuseId,
+                                     @RequestParam(name = DISC_APP_ID_QUERY_PARAM, required = false) Long appId,
+                                     @RequestParam(name = SUBMITTER_QUERY_PARAM, required = false) String submitter,
+                                     @RequestParam(name = EMAIL_QUERY_PARAM, required = false) String email,
+                                     @RequestParam(name = IP_QUERY_PARAM, required = false) String ip,
+                                     @RequestParam(name = SUBJECT_QUERY_PARAM, required = false) String subject,
+                                     @RequestParam(name = BODY_QUERY_PARAM, required = false) String body,
                                      AbuseViewModel abuseViewModel,
                                      Model model) {
         ReportedAbuse reportedAbuse = threadService.getReportedAbuse(abuseId);
@@ -63,18 +64,23 @@ public class AbuseController {
                 boolean isOwnerFound = false;
                 for (Application ownedApp : applicationService.getByOwnerId(user.getOwnerId())) {
                     if (ownedApp.getId().equals(reportedAbuse.getApplicationId())) {
-                        log.info("User: " + userEmail + " has deleted reported abuse: " + reportedAbuse.toString());
-                        threadService.deleteReportedAbuse(reportedAbuse.getId());
-                        abuseViewModel.setInfoMessage("Removed abuse record from database.");
-                        isOwnerFound = true;
-                        break;
+                        //make sure it's a user account or the system account that matches the app id.
+                        if (user.getIsUserAccount() || user.getEmail().equals(ownedApp.getId().toString())) {
+                            log.info("User: " + userEmail + " has deleted reported abuse id: " + reportedAbuse.getId());
+                            threadService.deleteReportedAbuse(reportedAbuse.getId());
+                            abuseViewModel.setInfoMessage("Removed abuse record from database.");
+                            isOwnerFound = true;
+                            break;
+                        }
                     }
                 }
                 if (!isOwnerFound) {
                     abuseViewModel.setErrorMessage("You cannot delete entries that you do not own.");
+                    log.warn("User: " + userEmail + " attempted to delete reported abuse id: " + reportedAbuse.getId()
+                            + " but they do not own it. Action has been blocked.");
                 }
             } else if (isRootAdmin) {
-                log.info("User: " + userEmail + " (ROOT) has deleted reported abuse: " + reportedAbuse.toString());
+                log.info("ROOT User has deleted reported abuse id: " + reportedAbuse.getId());
                 threadService.deleteReportedAbuse(reportedAbuse.getId());
                 abuseViewModel.setInfoMessage("Removed abuse record from database.");
             } else {
@@ -89,12 +95,12 @@ public class AbuseController {
     }
 
     @GetMapping(ABUSE_VIEW)
-    public ModelAndView getAbuseView(@RequestParam(name = "id", required = false) Long appId,
-                                     @RequestParam(name = "submitter", required = false) String submitter,
-                                     @RequestParam(name = "email", required = false) String email,
-                                     @RequestParam(name = "ip", required = false) String ip,
-                                     @RequestParam(name = "subject", required = false) String subject,
-                                     @RequestParam(name = "body", required = false) String body,
+    public ModelAndView getAbuseView(@RequestParam(name = ID_QUERY_PARAM, required = false) Long appId,
+                                     @RequestParam(name = SUBMITTER_QUERY_PARAM, required = false) String submitter,
+                                     @RequestParam(name = EMAIL_QUERY_PARAM, required = false) String email,
+                                     @RequestParam(name = IP_QUERY_PARAM, required = false) String ip,
+                                     @RequestParam(name = SUBJECT_QUERY_PARAM, required = false) String subject,
+                                     @RequestParam(name = BODY_QUERY_PARAM, required = false) String body,
                                      AbuseViewModel abuseViewModel,
                                      Model model) {
 
@@ -110,13 +116,22 @@ public class AbuseController {
         List<ReportedAbuse> reportedAbuses = threadService.searchForReportedAbuse(appId, submitter, email, ip, subject, body);
 
         //get list of apps owned by logged in user.
-        List<Application> ownedApps = null;
+        List<Long> ownedAppIds = new ArrayList<>();
         String userEmail = accountHelper.getLoggedInEmail();
         DiscAppUser user = userDetailsService.getByEmail(userEmail);
-        if (user != null && user.getOwnerId() != null) {
-            ownedApps = applicationService.getByOwnerId(user.getOwnerId());
 
+        //generate list of owned applications that user can delete records for
+        if (user != null && user.getOwnerId() != null) {
+            List<Application> ownedApps = applicationService.getByOwnerId(user.getOwnerId());
+            for (Application ownedApp : ownedApps) {
+                //only add matching app id for system accounts.
+                if (user.getIsUserAccount() || user.getEmail().equals(ownedApp.getId().toString())) {
+                    ownedAppIds.add(ownedApp.getId());
+                }
+            }
         }
+
+        boolean isRoot = accountHelper.isRootAdminAccount();
 
         for (ReportedAbuse reportedAbuse : reportedAbuses) {
             if (reportedAbuse.getThread() != null) {
@@ -124,38 +139,17 @@ public class AbuseController {
                 //if user owns app that abuse is reported for, flag it as deletable.
                 boolean isDeletable = false;
                 String deleteQueryParam = "";
-                if (ownedApps != null) {
-                    for (Application app : ownedApps) {
-                        if (reportedAbuse.getApplicationId().equals(app.getId())) {
-                            isDeletable = true;
-                            String appIdVal = "";
-                            if (appId != null) {
-                                appIdVal = appId.toString();
-                            }
 
-                            //TODO : store these strings somewhere.
-                            deleteQueryParam = "?abuseId=" + reportedAbuse.getId()
-                                    + "&discId=" + appIdVal
-                                    + "&submitter=" + abuseViewModel.getSubmitter()
-                                    + "&email=" + abuseViewModel.getEmail()
-                                    + "&ip=" + abuseViewModel.getIpAddress()
-                                    + "&subject=" + abuseViewModel.getSubject()
-                                    + "&body=" + abuseViewModel.getMessage();
-                            break;
-                        }
-                    }
-                }
-
-                //if logged in as site root admin, all entries should be deletable.
-                if (accountHelper.isRootAdminAccount()) {
+                if (isRoot || ownedAppIds.contains(reportedAbuse.getApplicationId())) {
                     isDeletable = true;
-                    deleteQueryParam = "?abuseId=" + reportedAbuse.getId()
-                            + "&discId=" + reportedAbuse.getApplicationId()
-                            + "&submitter=" + abuseViewModel.getSubmitter()
-                            + "&email=" + abuseViewModel.getEmail()
-                            + "&ip=" + abuseViewModel.getIpAddress()
-                            + "&subject=" + abuseViewModel.getSubject()
-                            + "&body=" + abuseViewModel.getMessage();
+
+                    deleteQueryParam = "?" + ABUSE_ID_QUERY_PARAM + "=" + reportedAbuse.getId()
+                            + "&" + DISC_APP_ID_QUERY_PARAM + "=" + reportedAbuse.getApplicationId()
+                            + "&" + SUBMITTER_QUERY_PARAM + "=" + abuseViewModel.getSubmitter()
+                            + "&" + EMAIL_QUERY_PARAM + "=" + abuseViewModel.getEmail()
+                            + "&" + IP_QUERY_PARAM + "=" + abuseViewModel.getIpAddress()
+                            + "&" + SUBJECT_QUERY_PARAM + "=" + abuseViewModel.getSubject()
+                            + "&" + BODY_QUERY_PARAM + "=" + abuseViewModel.getMessage();
                 }
 
                 AbuseViewModel.ReportedThread reportedThread = new AbuseViewModel.ReportedThread(
@@ -187,7 +181,7 @@ public class AbuseController {
     }
 
     @GetMapping(ABUSE_SEARCH_VIEW)
-    public ModelAndView getAbuseSearchView(@RequestParam(name = "articleId") long threadId,
+    public ModelAndView getAbuseSearchView(@RequestParam(name = ARTICLE_ID_QUERY_PARAM) long threadId,
                                            AbuseViewModel abuseViewModel,
                                            Model model) {
          Thread thread = threadService.getThreadById(threadId);
